refactor(gql): moved geometry utility functions from `schema.rs` to `geometry.rs`
