+++ crates/core/src/gql/schema.rs
@@ -31,6 +31,7 @@ use super::error::{resolver_error, GqlError};
 use super::ext::ValidatorExt;
 use crate::gql::error::{internal_error, schema_error, type_error};
 use crate::gql::ext::{NamedContainer, TryFromExt, TryIntoExt};
+use crate::gql::geometry;
 use crate::gql::utils::GqlValueUtils;
 use crate::kvs::LockType;
 use crate::kvs::TransactionType;
@@ -352,7 +353,7 @@ pub fn sql_value_to_gql_value(v: SqlValue) -> Result<GqlValue, GqlError> {
             Geometry::Line(line) => GqlValue::Object(
                 [
                     (Name::new("type"), GqlValue::String("LINE_STRING".to_string())),
-                    (Name::new("coordinates"), coord_collection_to_list(line)?),
+                    (Name::new("coordinates"), geometry::coord_collection_to_list(line)?),
                 ]
                     .into(),
             ),
@@ -364,7 +365,7 @@ pub fn sql_value_to_gql_value(v: SqlValue) -> Result<GqlValue, GqlError> {
                         GqlValue::List(
                             multiline
                                 .into_iter()
-                                .map(coord_collection_to_list)
+                                .map(geometry::coord_collection_to_list)
                                 .collect::<Result<Vec<_>, GqlError>>()?,
                         ),
                     ),
@@ -376,7 +377,7 @@ pub fn sql_value_to_gql_value(v: SqlValue) -> Result<GqlValue, GqlError> {
                     (Name::new("type"), GqlValue::String("MULTI_POINT".to_string())),
                     (
                         Name::new("coordinates"),
-                        coord_collection_to_list(multipoint.into_iter().map(|p| p.0))?,
+                        geometry::coord_collection_to_list(multipoint.into_iter().map(|p| p.0))?,
                     ),
                 ]
                     .into(),
@@ -384,7 +385,7 @@ pub fn sql_value_to_gql_value(v: SqlValue) -> Result<GqlValue, GqlError> {
             Geometry::Polygon(polygon) => GqlValue::Object(
                 [
                     (Name::new("type"), GqlValue::String("POLYGON".to_string())),
-                    (Name::new("coordinates"), polygon_to_list(&polygon)?),
+                    (Name::new("coordinates"), geometry::polygon_to_list(&polygon)?),
                 ]
                     .into(),
             ),
@@ -396,7 +397,7 @@ pub fn sql_value_to_gql_value(v: SqlValue) -> Result<GqlValue, GqlError> {
                         GqlValue::List(
                             multipolygon
                                 .iter()
-                                .map(polygon_to_list)
+                                .map(geometry::polygon_to_list)
                                 .collect::<Result<Vec<_>, _>>()?,
                         ),
                     ),
@@ -472,12 +473,12 @@ Result<TypeRef,
             0 => TypeRef::named("Geometry"),
             1 => {
                 let name = g.into_iter().next().expect("checked that length is 1");
-                TypeRef::named(geometry_kind_name_to_type_name(&name)?)
+                TypeRef::named(geometry::geometry_kind_name_to_type_name(&name)?)
             }
             _ => {
                 let geo_types = g
                     .iter()
-                    .map(|n| geometry_kind_name_to_type_name(n).map(TypeRef::named))
+                    .map(|n| geometry::geometry_kind_name_to_type_name(n).map(TypeRef::named))
                     .collect::<Result<Vec<_>, GqlError>>()?;
 
                 let union_name = format!("{}Union", ty_name);
@@ -629,6 +630,7 @@ macro_rules! any_try_kinds {
 }
 
 pub fn gql_to_sql_kind(val: &GqlValue, kind: Kind) -> Result<SqlValue, GqlError> {
+    use crate::gql::geometry;
     use crate::syn;
     match kind {
         Kind::Any => match val {
@@ -818,13 +820,13 @@ pub fn gql_to_sql_kind(val: &GqlValue, kind: Kind) -> Result<SqlValue, GqlError>
                     GqlValue::String(acutal_t) => {
                         let mut included = false;
                         for ty in ts {
-                            if geometry_kind_name_to_type_name(ty)? == acutal_t {
+                            if geometry::geometry_kind_name_to_type_name(ty)? == acutal_t {
                                 included = true;
                                 break;
                             }
                         }
                         if included {
-                            extract_geometry(map)
+                            geometry::extract_geometry(map)
                                 .map(SqlValue::Geometry)
                                 .ok_or_else(|| type_error(kind, val))
                         } else {
@@ -914,132 +916,5 @@ pub fn gql_to_sql_kind(val: &GqlValue, kind: Kind) -> Result<SqlValue, GqlError>
         Kind::Regex => Err(resolver_error("Regexes are not yet supported")),
         Kind::References(_, _) => Err(resolver_error("Cannot convert value into references")),
         Kind::File(_) => Err(resolver_error("Files are not yet supported")),
-    }
-}
-
-fn coord_to_list(coord: Coord<f64>) -> Result<GqlValue, GqlError> {
-    Ok(GqlValue::List(
-        <[f64; 2]>::from(coord)
-            .into_iter()
-            .map(GqlValue::try_fromx)
-            .collect::<Result<Vec<_>, _>>()?,
-    ))
-}
-
-fn coord_collection_to_list(
-    coord_collection: impl IntoIterator<Item=Coord<f64>>,
-) -> Result<GqlValue, GqlError> {
-    Ok(GqlValue::List(
-        coord_collection.into_iter().map(coord_to_list).collect::<Result<Vec<_>, _>>()?,
-    ))
-}
-
-fn polygon_to_list(polygon: &Polygon) -> Result<GqlValue, GqlError> {
-    Ok(GqlValue::List(
-        [polygon.exterior()]
-            .into_iter()
-            .chain(polygon.interiors().iter())
-            .cloned()
-            .map(coord_collection_to_list)
-            .collect::<Result<_, _>>()?,
-    ))
-}
-
-fn geometry_kind_name_to_type_name(name: &str) -> Result<&'static str, GqlError> {
-    match name {
-        "point" => Ok("GeometryPoint"),
-        "line" => Ok("GeometryLineString"),
-        "polygon" => Ok("GeometryPolygon"),
-        "multipoint" => Ok("GeometryMultiPoint"),
-        "multiline" => Ok("GeometryMultiLineString"),
-        "multipolygon" => Ok("GeometryMultiPolygon"),
-        "collection" => Ok("GeometryCollection"),
-        _ => Err(internal_error("expected valid geometry name")),
-    }
-}
-
-fn extract_coord(arr: &[GqlValue]) -> Option<Coord> {
-    match arr {
-        [GqlValue::Number(y), GqlValue::Number(x)] => Some(Coord {
-            x: x.as_f64()?,
-            y: y.as_f64()?,
-        }),
-        _ => None,
-    }
-}
-
-fn extract_coord_list(arr: &[GqlValue]) -> Option<Vec<Coord>> {
-    arr.iter()
-        .map(|c| match c {
-            GqlValue::List(c) => extract_coord(c),
-            _ => None,
-        })
-        .collect()
-}
-
-fn extract_coord_list_list(arr: &[GqlValue]) -> Option<Vec<Vec<Coord>>> {
-    arr.iter()
-        .map(|c| match c {
-            GqlValue::List(c) => extract_coord_list(c),
-            _ => None,
-        })
-        .collect()
-}
-
-fn extract_polygon(arr: &[GqlValue]) -> Option<Polygon> {
-    let mut line_strings = extract_coord_list_list(arr)?.into_iter().map(LineString);
-    let exterior = line_strings.next()?;
-    let interior = line_strings.collect();
-    Some(Polygon::new(exterior, interior))
-}
-
-fn extract_polygon_list(arr: &[GqlValue]) -> Option<Vec<Polygon>> {
-    arr.iter()
-        .map(|c| match c {
-            GqlValue::List(c) => extract_polygon(c),
-            _ => None,
-        })
-        .collect()
-}
-
-fn extract_geometry(map: &IndexMap<Name, GqlValue>) -> Option<Geometry> {
-    let ty = match map.get("type") {
-        Some(GqlValue::String(ty)) => Some(ty),
-        _ => None,
-    };
-
-    let coordinates = match map.get("coordinates") {
-        Some(GqlValue::List(cs)) => Some(cs.as_slice()),
-        _ => None,
-    };
-
-    let geometries = match map.get("geometries") {
-        Some(GqlValue::List(cs)) => Some(cs.as_slice()),
-        _ => None,
-    };
-
-    match ty?.as_str() {
-        "GeometryPoint" => Some(Geometry::Point(Point(extract_coord(coordinates?).unwrap()))),
-        "GeometryLineString" => Some(Geometry::Line(LineString(extract_coord_list(coordinates?)?))),
-        "GeometryPolygon" => Some(Geometry::Polygon(extract_polygon(coordinates?)?)),
-        "GeometryMultiPoint" => Some(Geometry::MultiPoint(MultiPoint(
-            extract_coord_list(&coordinates?)?.into_iter().map(Point).collect(),
-        ))),
-        "GeometryMultiLineString" => Some(Geometry::MultiLine(MultiLineString(
-            extract_coord_list_list(&coordinates?)?.into_iter().map(LineString).collect(),
-        ))),
-        "GeometryMultiPolygon" => {
-            Some(Geometry::MultiPolygon(MultiPolygon(extract_polygon_list(&coordinates?)?)))
-        }
-        "GeometryCollection" => Some(Geometry::Collection(
-            geometries?
-                .iter()
-                .map(|g| match g {
-                    GqlValue::Object(inner_map) => extract_geometry(inner_map),
-                    _ => None,
-                })
-                .collect::<Option<_>>()?,
-        )),
-        _ => None,
     }
 }
